style: rename GetKubeadmConfig -> GetRuntimeConfig

# Conflicts:
#	pkg/guest/guest.go
