+++ pkg/clusterfile/pre_process.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/labring/sealos/pkg/constants"
 	"github.com/labring/sealos/pkg/runtime/decode"
-	"github.com/labring/sealos/pkg/runtime/types"
+	"github.com/labring/sealos/pkg/runtime/kubernetes/types"
 	"github.com/labring/sealos/pkg/template"
 	v2 "github.com/labring/sealos/pkg/types/v1beta1"
 	fileutil "github.com/labring/sealos/pkg/utils/file"
@@ -36,10 +36,6 @@ var ErrClusterFileNotExists = errors.New("the cluster file is not exist")
 
 type PreProcessor interface {
 	Process() error
-}
-
-func NewPreProcessor(path string) PreProcessor {
-	return &ClusterFile{path: path}
 }
 
 func (c *ClusterFile) Process() (err error) {
@@ -124,7 +120,7 @@ func (c *ClusterFile) loadRenderValues() (map[string]interface{}, error) {
 
 func (c *ClusterFile) decode(data []byte) error {
 	for _, fn := range []func([]byte) error{
-		c.DecodeCluster, c.DecodeConfigs, c.DecodeKubeadmConfig,
+		c.DecodeCluster, c.DecodeConfigs, c.DecodeRuntimeConfig,
 	} {
 		if err := fn(data); err != nil && err != ErrTypeNotFound {
 			return err
@@ -158,14 +154,15 @@ func (c *ClusterFile) DecodeConfigs(data []byte) error {
 	return nil
 }
 
-func (c *ClusterFile) DecodeKubeadmConfig(data []byte) error {
-	kubeadmConfig, err := types.LoadKubeadmConfigs(string(data), c.setDefaults, decode.CRDFromString)
+func (c *ClusterFile) DecodeRuntimeConfig(data []byte) error {
+	// TODO: decode from k3s struct
+	config, err := types.LoadKubeadmConfigs(string(data), c.setDefaults, decode.CRDFromString)
 	if err != nil {
 		return err
 	}
-	if kubeadmConfig == nil {
+	if config == nil {
 		return ErrTypeNotFound
 	}
-	c.KubeConfig = kubeadmConfig
+	c.KubeConfig = config
 	return nil
 }
